Support new CMake and Gauss, and misc. improvements

New features:

- Use a subdirectory per platform for ccache
- Clone PARAM packages in a PARAM directory
- Support for new Gaudi cmake (distcc disabled for new-style projects)
- Support building Geant4 and Gauss (:warning: needs https://gitlab.cern.ch/lhcb-core/LbDevTools/-/merge_requests/163, see README on how to use a custom LbDevTools)

Improvements:

- VSCode: Match flake8 arguments in lhcb/Moore!712
- VSCode: Improve C/C++ extension configuration
- VSCode: Suppress warning from reStructuredText ext
- Explain how to use custom LbDevTools and lcg-toolchains
- Explain how to use Detector (and DD4hep)
- Document how to turn off the functor cache
- Update known issues in README
- Create TMPDIR if missing

Internal changes:

- Update to latest stable LbEnv (1059)
- Add LHCb/configure to CI add tests
- Add a CI test job for python 3
- Use default LHCb docker image and LbEnv Python 3
- Speed up project.mk a bit
- Set CMAKE_PREFIX_PATH in make.sh
- Temporary fix for Python 3 lookup

See merge request rmatev/lb-stack-setup!25

+++ project.mk
@@ -44,13 +44,11 @@
 # note that DIR will end with a /
 DIR := $(dir $(lastword $(MAKEFILE_LIST)))
 PROJECT := $(notdir $(CURDIR))
-_output_path := $(shell "$(DIR)config.py" outputPath)
-_dummy := $(shell mkdir -p "$(_output_path)" && printenv | sort > "$(_output_path)/project.mk.env")
 
 # settings
 CMAKE := cmake
 CTEST := ctest
-NINJA := $(shell which ninja 2> /dev/null)
+NINJA := ninja
 
 # modified by RM
 ifeq ($(findstring CMAKE_TOOLCHAIN_FILE,$(CMAKEFLAGS)),)
@@ -154,7 +152,7 @@ ifeq ($(VERBOSE),)
 # (emulate the default CMake install target)
 install: patch-python-ns
 	cd $(BUILDDIR) && $(CMAKE) -P cmake_install.cmake | grep -v "^-- Up-to-date:"
-	cp $(BUILDDIR)/config/$(PROJECT)-build.xenv $(INSTALLDIR)/$(PROJECT).xenv
+	test -f $(BUILDDIR)/config/$(PROJECT)-build.xenv && cp $(BUILDDIR)/config/$(PROJECT)-build.xenv $(INSTALLDIR)/$(PROJECT).xenv || true
 endif
 
 # ensure that the target are always passed to the CMake Makefile
